## What is Github Git is a **version control** software. It maintain different versions of our code/projects as well a history of out changes.
## What is a commit
In git a **commit** is a checkpoint of the changes we made in our project. It is important to put clear **commit** messages, so you know what changes had occured there.

## What is a branch
A branch is a different version of our code that comes from a certain commit. Changes that occur in seperate branches don't effect main branch or other branches. They are independent until they are merged into other branches.